Renames UWP files.
