+++ src/HtmlLabel/UWP/Behavior.cs
@@ -1,24 +1,29 @@
-﻿using System;
-using Windows.UI.Xaml;
+﻿using Windows.UI.Xaml;
 using LabelHtml.Forms.Plugin.Abstractions;
 using Xamarin.Forms.Platform.UWP;
 using LabelHtml.Forms.Plugin.UWP;
+using Microsoft.Xaml.Interactivity;
 
 [assembly: ExportRenderer(typeof(HtmlLabel), typeof(HtmlLabelRenderer))]
 // ReSharper disable once CheckNamespace
 namespace LabelHtml.Forms.Plugin.UWP
 {
-    internal abstract class Behavior<T> : Behavior where T : DependencyObject
+    internal abstract class Behavior : DependencyObject, IBehavior
 	{
-		protected new T AssociatedObject => base.AssociatedObject as T;
-
-		protected override void OnAttached()
+		public void Attach(DependencyObject associatedObject)
 		{
-			base.OnAttached();
-			if (AssociatedObject == null)
-            {
-                throw new InvalidOperationException("AssociatedObject is not of the right type");
-            }
-        }
+			AssociatedObject = associatedObject;
+			OnAttached();
+		}
+
+		public void Detach() => OnDetaching();
+
+		protected virtual void OnAttached() { }
+
+		protected virtual void OnDetaching() { }
+
+		protected DependencyObject AssociatedObject { get; set; }
+
+		DependencyObject IBehavior.AssociatedObject => AssociatedObject;
 	}
 }
